# Support large number of projections in CursorProcessor