… core/trino-main/src/main/java/io/trino/sql/gen/BodyCompiler.java
… core/trino-main/src/test/java/io/trino/metadata/TestingFunctionResolution.java
… core/trino-main/src/test/java/io/trino/operator/BenchmarkScanFilterAndProjectOperator.java
… core/trino-main/src/test/java/io/trino/operator/TestColumnarPageProcessor.java
… core/trino-main/src/test/java/io/trino/operator/TestScanFilterAndProjectOperator.java
… core/trino-main/src/test/java/io/trino/sql/gen/BenchmarkInCodeGenerator.java
… core/trino-main/src/test/java/io/trino/sql/gen/BenchmarkPageProcessor2.java
… core/trino-main/src/test/java/io/trino/sql/gen/TestCursorProcessorCompiler.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestOrcPageSourceMemoryTracking.java