<!-- Thank you for submitting a pull request! Find more information 
at https://trino.io/development/process.html, 
at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md 
and contact us on #core-dev in Slack. -->
<!-- Provide an overview for maintainers and reviewers. -->
## Description

### Problem
Queries with a large number of projections fail with a `MethodTooLargeException` (e.g., Method too large: `io/trino/$gen/CursorProcessor_20250217_061519_14484.process`).

Example failing query:
```sql
SELECT  
    nationkey,  
    nationkey + 1,  
    nationkey + 2,  
    nationkey + 3,  
    nationkey + 4,  
    ...  
    nationkey + 5000  
FROM tpch.sf1.nation;
```

After applying the proposed solution below, this query can run successfully with 100_000+ projections.

### Proposed Solution

Instead of generating the `process` method like this:

```java
public CursorProcessorOutput process(ConnectorSession session, DriverYieldSignal yieldSignal, RecordCursor cursor, PageBuilder pageBuilder) {
    int completedPositions = 0;
    boolean finished;
    
    for (finished = false; true && !pageBuilder.isFull() && !yieldSignal.isSet(); ++completedPositions) {
        if (!cursor.advanceNextPosition()) {
            finished = true;
            break;
        }

        if (this.filter(session, cursor)) {
            pageBuilder.declarePosition();
            this.project_0(session, cursor, pageBuilder.getBlockBuilder(0));
            this.project_1(session, cursor, pageBuilder.getBlockBuilder(1));
            this.project_2(session, cursor, pageBuilder.getBlockBuilder(2));
            ...
        }
    }

    return new CursorProcessorOutput(completedPositions, finished);
}
```

This PR proposes splitting the logic into multiple classes:

```java
public final class CursorProcessor_20250217_075231_713 implements CursorProcessor {
    private final CursorProcessorChunk_20250217_075231_699 chunk_0 = new CursorProcessorChunk_20250217_075231_699();
    private final CursorProcessorChunk_20250217_075231_709 chunk_1 = new CursorProcessorChunk_20250217_075231_709();
    private final CursorProcessorChunk_20250217_075231_712 chunk_2 = new CursorProcessorChunk_20250217_075231_712();
    ...

    public CursorProcessorOutput process(ConnectorSession session, DriverYieldSignal yieldSignal, RecordCursor cursor, PageBuilder pageBuilder) {
        int completedPositions = 0;

        boolean finished;
        for (finished = false; true && !pageBuilder.isFull() && !yieldSignal.isSet(); ++completedPositions) {
            if (!cursor.advanceNextPosition()) {
                finished = true;
                break;
            }

            if (this.filter(session, cursor)) {
                pageBuilder.declarePosition();
                this.chunk_0.project(session, cursor, pageBuilder);
                this.chunk_1.project(session, cursor, pageBuilder);
                this.chunk_2.project(session, cursor, pageBuilder);
                ...
            }
        }

        return new CursorProcessorOutput(completedPositions, finished);
    }

    ...
}
```

Each chunk handles a fixed number of projections (e.g., 100):
```java
public final class CursorProcessorChunk_20250217_075231_699 {
    public void project(ConnectorSession session, RecordCursor cursor, PageBuilder pageBuilder) {
        this.project_0(session, cursor, pageBuilder.getBlockBuilder(0));
        this.project_1(session, cursor, pageBuilder.getBlockBuilder(1));
        this.project_2(session, cursor, pageBuilder.getBlockBuilder(2));
        this.project_3(session, cursor, pageBuilder.getBlockBuilder(3));
        ...
    }
    ...
}
```

#### Why not batch projections into smaller methods in CursorProcessor?
With this approach, we would quickly hit another JVM limit: the number of entries in the constant pool.

#### Can we avoid creating `CursorProcessorChunk_*` instances and instead make `CursorProcessorChunk_*::process` static?
No, because some row expressions rely on `__cachedInstance` fields being declared in the class that evaluates them.

### Benchmarks
The following `BenchmarkPageProcessor2` results were obtained on a MacBook Pro M3 laptop.

I haven’t yet investigated the reason for improvements with a small number of projections (2-32). However, for a large number of projections, the improvement is expected based on compilation logs - previously, `CursorProcessor::process` was not JIT-compiled because it was too large.

![image](https://github.com/user-attachments/assets/578c5933-324f-4e14-9609-75f910baf124)

#### Before
```
Benchmark                            (columnCount)  (dictionaryBlocks)   (type)  Mode  Cnt         Score        Error  Units
BenchmarkPageProcessor2.rowOriented              2               false  varchar  avgt   50     50160.535 ±     594.272  ns/op
BenchmarkPageProcessor2.rowOriented              2               false   bigint  avgt   50      7810.172 ±     571.491  ns/op
BenchmarkPageProcessor2.rowOriented              2                true  varchar  avgt   50     51470.135 ±     315.766  ns/op
BenchmarkPageProcessor2.rowOriented              2                true   bigint  avgt   50     10378.700 ±     122.560  ns/op
BenchmarkPageProcessor2.rowOriented              4               false  varchar  avgt   50     83045.705 ±    1268.451  ns/op
BenchmarkPageProcessor2.rowOriented              4               false   bigint  avgt   50     13823.784 ±     273.724  ns/op
BenchmarkPageProcessor2.rowOriented              4                true  varchar  avgt   50     87986.951 ±     569.324  ns/op
BenchmarkPageProcessor2.rowOriented              4                true   bigint  avgt   50     17700.610 ±     101.524  ns/op
BenchmarkPageProcessor2.rowOriented              8               false  varchar  avgt   50    118756.548 ±     819.084  ns/op
BenchmarkPageProcessor2.rowOriented              8               false   bigint  avgt   50     27241.108 ±     496.276  ns/op
BenchmarkPageProcessor2.rowOriented              8                true  varchar  avgt   50    128909.003 ±    1362.939  ns/op
BenchmarkPageProcessor2.rowOriented              8                true   bigint  avgt   50     32884.000 ±     526.453  ns/op
BenchmarkPageProcessor2.rowOriented             16               false  varchar  avgt   50    195338.961 ±    1356.504  ns/op
BenchmarkPageProcessor2.rowOriented             16               false   bigint  avgt   50     53924.685 ±     516.367  ns/op
BenchmarkPageProcessor2.rowOriented             16                true  varchar  avgt   50    221104.584 ±    1716.688  ns/op
BenchmarkPageProcessor2.rowOriented             16                true   bigint  avgt   50     62181.934 ±     780.327  ns/op
BenchmarkPageProcessor2.rowOriented             32               false  varchar  avgt   50    366533.046 ±    4294.019  ns/op
BenchmarkPageProcessor2.rowOriented             32               false   bigint  avgt   50    107821.145 ±     583.700  ns/op
BenchmarkPageProcessor2.rowOriented             32                true  varchar  avgt   50    415592.071 ±    6571.041  ns/op
BenchmarkPageProcessor2.rowOriented             32                true   bigint  avgt   50    121137.366 ±     892.216  ns/op
BenchmarkPageProcessor2.rowOriented           1024               false  varchar  avgt   50  11021790.298 ± 1420506.871  ns/op
BenchmarkPageProcessor2.rowOriented           1024               false   bigint  avgt   50   6952926.742 ±  111941.282  ns/op
BenchmarkPageProcessor2.rowOriented           1024                true  varchar  avgt   50  12413085.100 ± 1684711.540  ns/op
BenchmarkPageProcessor2.rowOriented           1024                true   bigint  avgt   50   7188688.105 ±  146991.396  ns/op
BenchmarkPageProcessor2.rowOriented           4000               false  varchar  avgt   50  56729491.760 ± 2916380.205  ns/op
BenchmarkPageProcessor2.rowOriented           4000               false   bigint  avgt   50  16900532.808 ±  933993.951  ns/op
BenchmarkPageProcessor2.rowOriented           4000                true  varchar  avgt   50  59424635.399 ± 3984873.077  ns/op
BenchmarkPageProcessor2.rowOriented           4000                true   bigint  avgt   50  17508239.933 ±  930771.108  ns/op
```

#### This PR

```
Benchmark                            (columnCount)  (dictionaryBlocks)   (type)  Mode  Cnt         Score         Error  Units
BenchmarkPageProcessor2.rowOriented              2               false  varchar  avgt   50     46220.098 ±     881.331  ns/op
BenchmarkPageProcessor2.rowOriented              2               false   bigint  avgt   50      7946.137 ±      19.637  ns/op
BenchmarkPageProcessor2.rowOriented              2                true  varchar  avgt   50     45184.346 ±    1056.650  ns/op
BenchmarkPageProcessor2.rowOriented              2                true   bigint  avgt   50      9555.191 ±      14.417  ns/op
BenchmarkPageProcessor2.rowOriented              4               false  varchar  avgt   50     79137.921 ±    1078.074  ns/op
BenchmarkPageProcessor2.rowOriented              4               false   bigint  avgt   50     13617.031 ±      73.185  ns/op
BenchmarkPageProcessor2.rowOriented              4                true  varchar  avgt   50     82802.368 ±    2362.382  ns/op
BenchmarkPageProcessor2.rowOriented              4                true   bigint  avgt   50     16091.008 ±     123.314  ns/op
BenchmarkPageProcessor2.rowOriented              8               false  varchar  avgt   50    119164.189 ±    1756.264  ns/op
BenchmarkPageProcessor2.rowOriented              8               false   bigint  avgt   50     25650.300 ±     638.566  ns/op
BenchmarkPageProcessor2.rowOriented              8                true  varchar  avgt   50    128802.703 ±    1624.458  ns/op
BenchmarkPageProcessor2.rowOriented              8                true   bigint  avgt   50     31939.751 ±     658.227  ns/op
BenchmarkPageProcessor2.rowOriented             16               false  varchar  avgt   50    184912.270 ±    1155.579  ns/op
BenchmarkPageProcessor2.rowOriented             16               false   bigint  avgt   50     47894.594 ±     120.073  ns/op
BenchmarkPageProcessor2.rowOriented             16                true  varchar  avgt   50    210042.296 ±     637.544  ns/op
BenchmarkPageProcessor2.rowOriented             16                true   bigint  avgt   50     59369.979 ±     111.791  ns/op
BenchmarkPageProcessor2.rowOriented             32               false  varchar  avgt   50    361427.695 ±    3990.169  ns/op
BenchmarkPageProcessor2.rowOriented             32               false   bigint  avgt   50     98585.665 ±     292.521  ns/op
BenchmarkPageProcessor2.rowOriented             32                true  varchar  avgt   50    406628.732 ±    5439.313  ns/op
BenchmarkPageProcessor2.rowOriented             32                true   bigint  avgt   50    118210.373 ±     347.720  ns/op
BenchmarkPageProcessor2.rowOriented           1024               false  varchar  avgt   50   4882329.085 ±  654439.458  ns/op
BenchmarkPageProcessor2.rowOriented           1024               false   bigint  avgt   50   1495624.500 ±   30682.770  ns/op
BenchmarkPageProcessor2.rowOriented           1024                true  varchar  avgt   50   7093766.811 ± 1440875.938  ns/op
BenchmarkPageProcessor2.rowOriented           1024                true   bigint  avgt   50   1994875.420 ±   48772.880  ns/op
BenchmarkPageProcessor2.rowOriented           4000               false  varchar  avgt   50  51464647.067 ± 2826170.499  ns/op
BenchmarkPageProcessor2.rowOriented           4000               false   bigint  avgt   50   6816984.051 ±  817155.683  ns/op
BenchmarkPageProcessor2.rowOriented           4000                true  varchar  avgt   50  51559848.524 ± 2003482.263  ns/op
BenchmarkPageProcessor2.rowOriented           4000                true   bigint  avgt   50   8925249.184 ±  746383.609  ns/op
```


<!-- Provide details that help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues



<!-- Mark the appropriate option with an (x). Propose a release note if you can.
More info at https://trino.io/development/process#release-note -->
## Release notes

( ) This is not user-visible or is docs only, and no release notes are required.
( ) Release notes are required. Please propose a release note for me.
(x) Release notes are required, with the following suggested text:

```markdown
## Section
* Add support for large number of projections. ({issue}`issuenumber`)
```