@losipiuk 2025-02-17
stupid question (i did not read the code yet). Why do we need `CursorProcessorChunk_*` classes - and not just extra methods in top-level `CursorProcessorChunk_*` class.