@wendigo 2025-02-17
@losipiuk I guess that there is a limit to the number of methods and class size