@piotrrzysko 2025-02-17
Quote from the PR description (it is long, sorry): 

> Why not batch projections into smaller methods in CursorProcessor?
With this approach, we would quickly hit another JVM limit: the number of entries in the constant pool.

The biggest contributor to the constant pool entries is the `invokedynamic` instructions. Each requires 2–3 entries in the pool, and we have at least two `invokedynamic` instructions per projection. By introducing multiple classes, I also split the constant pool.